refactor: modularize NFT data and table configuration into composables
